Store Interface

+++ src/main/java/com/example/gzone/StoreProfileController.java
@@ -1,8 +1,12 @@
 package com.example.gzone;
 
 import com.example.entity.MarketItem;
+import com.example.entity.Store;
+import com.example.entity.User;
+import com.example.service.Games;
 import com.example.service.MarketItems;
 import com.example.service.Stores;
+import com.example.service.Users;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -12,6 +16,7 @@ import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.AnchorPane;
+import javafx.scene.text.Text;
 import javafx.stage.Stage;
 
 import java.io.IOException;
@@ -38,18 +43,26 @@ public class StoreProfileController {
         public  AnchorPane profilepane;
         @FXML
         public Button deletebutton;
+        @FXML
+        private Text storenametext;
+        @FXML
+        private Text usernametext ;
 
         @FXML
         void AddToStore(ActionEvent event) throws IOException {
-            AnchorPane pane = FXMLLoader.load(getClass().getResource("AddItemForm.fxml"));
-            profilepane.getChildren().setAll(pane);
+                if (new Stores().findById(Id.store).getOwnerId().equals(Id.user)) {
+                        AnchorPane pane = FXMLLoader.load(getClass().getResource("AddItemForm.fxml"));
+                        profilepane.getChildren().setAll(pane);
+                }
         }
         @FXML
         void DeleteYourItem(ActionEvent event) throws IOException {
-                new MarketItems().deleteById(
-                        ((MarketItem) tbview.getSelectionModel().getSelectedItem()).getId()
-                );
-                ViewItem();
+                if (new Stores().findById(Id.store).getOwnerId().equals(Id.user)) {
+                        new MarketItems().deleteById(
+                                ((MarketItem) tbview.getSelectionModel().getSelectedItem()).getId()
+                        );
+                        ViewItem();
+                }
         }
 
         @FXML
@@ -65,7 +78,7 @@ public class StoreProfileController {
 
         @FXML
         void Store(ActionEvent event) throws IOException{
-                AnchorPane pane = FXMLLoader.load(getClass().getResource("Store2.fxml"));
+                AnchorPane pane = FXMLLoader.load(getClass().getResource("ViewStores.fxml"));
                 profilepane.getChildren().setAll(pane);
         }
 
@@ -83,20 +96,17 @@ public class StoreProfileController {
         void initialize(){
                 cltitle.setCellValueFactory(new PropertyValueFactory<>("title"));
                 tbview.getColumns().add(cltitle);
-
-                List<MarketItem> l = new MarketItems().findAll("`store_id`=" + 63);
-                tbview.getItems().clear();
-                for (MarketItem m : l) {
-                        tbview.getItems().add(m);
-
-                }
-                tbview.refresh();
+                Store s = new Stores().findById(Id.store);
+                User u = new Users().findById(s.getOwnerId());
+                storenametext.setText(s.getName());
+                usernametext.setText(u.getUsername());
+                ViewItem();
         }
 
         @FXML
         public void ViewItem() {
 
-                List<MarketItem> l = new MarketItems().findAll("`store_id`=" + 63);
+                List<MarketItem> l = new MarketItems().findAll("`store_id`=" + Id.store);
                 tbview.getItems().clear();
                 for (MarketItem m : l) {
                         tbview.getItems().add(m);
@@ -105,8 +115,12 @@ public class StoreProfileController {
                 tbview.refresh();
         }
         @FXML
-        void DeleteYourStore(ActionEvent event) {
-             //   new Stores().deleteById(id);
+        void DeleteYourStore(ActionEvent event)throws IOException {
+                if (new Stores().findById(Id.store).getOwnerId().equals(Id.user)) {
+                        new MarketItems().deleteByStoreId(Id.store);
+                        new Stores().deleteById(Id.store);
+                        Store(event);
+                }
         }
